+++ experimaestro/tests/test_tasks.py
@@ -7,9 +7,9 @@ import signal
 from experimaestro import *
 from experimaestro.scheduler import JobState
 
-from .utils import TemporaryDirectory, TemporaryExperiment
+from .utils import TemporaryDirectory, TemporaryExperiment, get_times
 
-from .tasks import *
+from .tasks.all import *
 from . import restart
 
 
@@ -49,7 +49,7 @@ def test_foreign_type():
     """When the argument real type is in an non imported module"""
     with TemporaryExperiment("foreign_type", maxwait=2):
         # Submit the tasks
-        from .tasks2 import ForeignClassB2
+        from .tasks.foreign import ForeignClassB2
 
         b = ForeignClassB2(x=1, y=2)
         a = ForeignTaskA(b=b).submit()
@@ -115,3 +115,20 @@ def test_configcache():
         task = CacheConfigTask(data=CacheConfig()).submit()
 
     assert task.__xpm__.job.wait() == JobState.DONE
+
+
+def test_subparams():
+    """Test sub-parameters that allow a main task to contain shared data
+
+    This is useful when e.g. training with a different number of epochs
+    """
+    from .tasks.subparams import Task
+
+    with TemporaryExperiment("subparam", maxwait=10) as xp:
+        task100 = Task(epoch=100, x=1).submit()
+        task200 = Task(epoch=200, x=1).submit()
+
+        xp.wait()
+
+        t100, t200 = get_times(task100), get_times(task200)
+        assert t100 > t200 or t200 > t100